requirements.txt
langchain_openai==1.1.6
numpy==2.3.5
pandas==2.3.3
python-dotenv==1.2.1
statsmodels==0.14.5
streamlit==1.51.0
